Rename project to gameoffe
