+++ web/index.html
@@ -23,13 +23,13 @@
   <!-- iOS meta tags & icons -->
   <meta name="mobile-web-app-capable" content="yes">
   <meta name="apple-mobile-web-app-status-bar-style" content="black">
-  <meta name="apple-mobile-web-app-title" content="gemeoffe">
+  <meta name="apple-mobile-web-app-title" content="gameoffe">
   <link rel="apple-touch-icon" href="icons/Icon-192.png">
 
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>gemeoffe</title>
+  <title>gameoffe</title>
   <link rel="manifest" href="manifest.json">
 </head>
 <body>
