+++ web/manifest.json
@@ -1,6 +1,6 @@
 {
-    "name": "gemeoffe",
-    "short_name": "gemeoffe",
+    "name": "gameoffe",
+    "short_name": "gameoffe",
     "start_url": ".",
     "display": "standalone",
     "background_color": "#0175C2",
