+++ windows/runner/Runner.rc
@@ -90,12 +90,12 @@ BEGIN
         BLOCK "040904e4"
         BEGIN
             VALUE "CompanyName", "com.example" "\0"
-            VALUE "FileDescription", "gemeoffe" "\0"
+            VALUE "FileDescription", "gameoffe" "\0"
             VALUE "FileVersion", VERSION_AS_STRING "\0"
-            VALUE "InternalName", "gemeoffe" "\0"
+            VALUE "InternalName", "gameoffe" "\0"
             VALUE "LegalCopyright", "Copyright (C) 2025 com.example. All rights reserved." "\0"
-            VALUE "OriginalFilename", "gemeoffe.exe" "\0"
-            VALUE "ProductName", "gemeoffe" "\0"
+            VALUE "OriginalFilename", "gameoffe.exe" "\0"
+            VALUE "ProductName", "gameoffe" "\0"
             VALUE "ProductVersion", VERSION_AS_STRING "\0"
         END
     END
